Update to current codebase

+++ code/game/objects/effects/decals/cleanable.dm
@@ -1,9 +1,12 @@
 /obj/effect/decal/cleanable
+	gender = PLURAL
+	layer = ABOVE_NORMAL_TURF_LAYER
 	var/list/random_icon_states = list()
 	var/blood_state = "" //I'm sorry but cleanable/blood code is ass, and so is blood_DNA
 	var/bloodiness = 0 //0-100, amount of blood in this decal, used for making footprints and affecting the alpha of bloody footprints
+	var/mergeable_decal = TRUE //when two of these are on a same tile or do we need to merge them into just one?
 
-/obj/effect/decal/cleanable/New()
+/obj/effect/decal/cleanable/Initialize(mapload, list/datum/disease/diseases)
 	if (random_icon_states && length(src.random_icon_states) > 0)
 		src.icon_state = pick(src.random_icon_states)
 	create_reagents(300)
@@ -11,24 +14,30 @@
 		for(var/obj/effect/decal/cleanable/C in src.loc)
 			if(C != src && C.type == src.type)
 				replace_decal(C)
-	..()
-
-
+	if(LAZYLEN(diseases))
+		var/list/datum/disease/diseases_to_add = list()
+		for(var/datum/disease/D in diseases)
+			if(D.spread_flags & VIRUS_SPREAD_CONTACT_FLUIDS)
+				diseases_to_add += D
+		if(LAZYLEN(diseases_to_add))
+			AddComponent(/datum/component/infective, diseases_to_add)
+	. = ..()
 
 /obj/effect/decal/cleanable/proc/replace_decal(obj/effect/decal/cleanable/C)
-	qdel(C)
+	if(mergeable_decal)
+		qdel(C)
 
-/obj/effect/decal/cleanable/attackby(obj/item/weapon/W, mob/user, params)
-	if(istype(W, /obj/item/weapon/reagent_containers/glass) || istype(W, /obj/item/weapon/reagent_containers/food/drinks))
+/obj/effect/decal/cleanable/attackby(obj/item/W, mob/user, params)
+	if(istype(W, /obj/item/reagent_containers/glass) || istype(W, /obj/item/reagent_containers/food/drinks))
 		if(src.reagents && W.reagents)
 			. = 1 //so the containers don't splash their content on the src while scooping.
 			if(!src.reagents.total_volume)
-				user << "<span class='notice'>[src] isn't thick enough to scoop up!</span>"
+				to_chat(user, "<span class='notice'>[src] isn't thick enough to scoop up!</span>")
 				return
 			if(W.reagents.total_volume >= W.reagents.maximum_volume)
-				user << "<span class='notice'>[W] is full!</span>"
+				to_chat(user, "<span class='notice'>[W] is full!</span>")
 				return
-			user << "<span class='notice'>You scoop up [src] into [W]!</span>"
+			to_chat(user, "<span class='notice'>You scoop up [src] into [W]!</span>")
 			reagents.trans_to(W, reagents.total_volume)
 			if(!reagents.total_volume) //scooped up all of it
 				qdel(src)
@@ -38,10 +47,10 @@
 			return
 		else
 			var/hotness = W.is_hot()
-			var/added_heat = (hotness / 100)
-			src.reagents.chem_temp = min(src.reagents.chem_temp + added_heat, hotness)
-			src.reagents.handle_reactions()
-			user << "<span class='notice'>You heat [src] with [W]!</span>"
+			reagents.expose_temperature(hotness)
+			to_chat(user, "<span class='notice'>You heat [name] with [W]!</span>")
+	else
+		return ..()
 
 /obj/effect/decal/cleanable/ex_act()
 	if(reagents)
@@ -49,16 +58,16 @@
 			R.on_ex_act()
 	..()
 
-/obj/effect/decal/cleanable/fire_act()
+/obj/effect/decal/cleanable/fire_act(exposed_temperature, exposed_volume)
 	if(reagents)
-		reagents.chem_temp += 30
-		reagents.handle_reactions()
+		reagents.expose_temperature(exposed_temperature)
 	..()
 
 
 //Add "bloodiness" of this blood's type, to the human's shoes
 //This is on /cleanable because fuck this ancient mess
 /obj/effect/decal/cleanable/Crossed(atom/movable/O)
+	..()
 	if(ishuman(O))
 		var/mob/living/carbon/human/H = O
 		if(H.shoes && blood_state && bloodiness)
